Move validators into new files.. is this better?
